Custom parameters file `parameters.md`
Moved copies of parameters to text file `parameters.md`
to make it easier to reproduce the files uploaded to BGG,
and settings for personal files to print.
The idea is to copy the settings in this file and paste them
into the R Markdown file temporarily for a specific output,
without having to manually move a bunch of other (input) files around.