Changes to reduce output file size (~10%)
Trim output images, and strip the color profiles.
The color profile is incorrect, and should probably be tagged in the pdf file once, rather than in every image.
strip() is probably accounting for most of the reduction in final file size.
Trim() is unreliable, resulting in possible differences in content 'height', so some changes to the \parbox code were needed to preserve vertical alignment.